Rename DaemonConnections to ProcessorRegistry
