--- lib/flight_scheduler/daemon_connections.rb
@@ -1,76 +1,0 @@
-#==============================================================================
-# Copyright (C) 2020-present Alces Flight Ltd.
-#
-# This file is part of FlightSchedulerController.
-#
-# This program and the accompanying materials are made available under
-# the terms of the Eclipse Public License 2.0 which is available at
-# <https://www.eclipse.org/legal/epl-2.0>, or alternative license
-# terms made available by Alces Flight Ltd - please direct inquiries
-# about licensing to licensing@alces-flight.com.
-#
-# FlightSchedulerController is distributed in the hope that it will be useful, but
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, EITHER EXPRESS OR
-# IMPLIED INCLUDING, WITHOUT LIMITATION, ANY WARRANTIES OR CONDITIONS
-# OF TITLE, NON-INFRINGEMENT, MERCHANTABILITY OR FITNESS FOR A
-# PARTICULAR PURPOSE. See the Eclipse Public License 2.0 for more
-# details.
-#
-# You should have received a copy of the Eclipse Public License 2.0
-# along with FlightSchedulerController. If not, see:
-#
-#  https://opensource.org/licenses/EPL-2.0
-#
-# For more information on FlightSchedulerController, please visit:
-# https://github.com/openflighthpc/flight-scheduler-controller
-#==============================================================================
-require 'concurrent'
-
-# TODO: Move into a FlightScheduler namespace or app dir
-class DaemonConnections
-  class DuplicateConnection < RuntimeError; end
-  class UnconnectedNode < RuntimeError ; end
-
-  def initialize
-    @connections = Concurrent::Hash.new
-  end
-
-  def add(node_name, processor)
-    if @connections[node_name]
-      raise DuplicateConnection, node_name
-    end
-    @connections[node_name] = processor
-    FlightScheduler.app.event_processor.node_connected
-  end
-
-  def remove(processor)
-    @connections.delete(processor.node_name)
-  end
-
-  def [](node_name)
-    @connections[node_name]
-  end
-
-  def connected_nodes
-    @connections.keys
-  end
-
-  # TODO: REMOVE ME! Replace with a dedicated processor
-  def connection_for(node_name)
-    processor = @connections[node_name]
-    raise UnconnectedNode, node_name if processor.nil?
-    processor.connection
-  end
-
-  # TODO: Cache the processors
-  def daemon_processor_for(node_name)
-    con = connection_for(node_name)
-    FlightScheduler::EventProcessor::DaemonProcessor.new(con, node_name)
-  end
-
-  # TODO: Cache the processors
-  def job_processor_for(node_name, job_id)
-    con = connection_for(node_name)
-    FlightScheduler::EventProcessor::JobProcessor.new(con, node_name, job_id)
-  end
-end
